ep03_Instant_triviabot/pyproject.toml
[project]
name = "instant-trivia-bot"
version = "0.1.0"
description = "Async streaming trivia bot with LangChain v0.3+"
authors = [
    {name = "consciouscode",email = "consciouscode.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.26",
    "langchain-openai==0.3.28",
    "openai==1.97.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "rich>=13.9.4"
]

[tool.poetry]
packages = [{include = "instant_trivia_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24.0"
black = "^24.0.0"
ruff = "^0.4"

[tool.poetry.scripts]
start = "instant_trivia_bot.chatbot:main"
parallel = "instant_trivia_bot.parallel_demo:run"
structured = "instant_trivia_bot.structured_output:run_structured_demo"